Schedule timecard review upon button click
There will be times when a timecard review cannot be automatically
triggered in response to an incoming message (for example if the
necessary fields could not be identified in the message). In this case,
the user will need to manually trigger a review by clicking the review
button in the communications view. This change ensured an outgoing
message is scheduled to be sent when the user clicks this button.