Fix a bug when inferring the job name
Currently, we are inferring the name from the most recent entry. We
should be choosing the job name from the first job in the prior day. This
is the job that will most likely match up with the job we assume in the
daily reminder (since the first job is what triggers the reminder).